change code

+++ lib/widgets/Note_ListTile.dart
@@ -3,6 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:noteapp/controller/note_cubit.dart';
 import 'package:noteapp/model/NoteModel.dart';
 import 'package:noteapp/views/Edit_Note.dart';
+import 'package:noteapp/views/home.dart';
 
 class NoteListtile extends StatelessWidget {
   const NoteListtile({super.key,required this.note,required this.index});
@@ -37,7 +38,40 @@ class NoteListtile extends StatelessWidget {
                 style: TextStyle(fontSize: 18, color: Colors.grey),
               ),
               trailing: IconButton(onPressed: (){
-                BlocProvider.of<NoteCubit>(context).DeleteNote(index);
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return new AlertDialog(
+                      title: Text("Delete Confirmation"),
+                      content: Text(
+                        "Are you sure you want to delete this item?",
+                      ),
+                      actions: [
+                        TextButton(
+                          onPressed: () => Navigator.pop(context), // Cancel
+                          child: Text("Cancel"),
+                        ),
+                        TextButton(
+                          onPressed: () async {
+                            await BlocProvider.of<NoteCubit>(context).DeleteNote(index);
+                            Navigator.pushAndRemoveUntil(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => Home(),
+                              ),
+                                  (route) => false,
+                            ); // Close dialog
+                          },
+                          child: Text(
+                            "Delete",
+                            style: TextStyle(color: Colors.red),
+                          ),
+                        ),
+                      ],
+                    );
+                  },
+                );
+
               }, icon: Icon(
                 Icons.delete,
                 size: 30,
